+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unrelease]
 - Fix S3 Bucket forbidden
+- Added #18389 - APM3.0 + Datahub3.0
 
 ## [3.4.2] - 2020-12-02
 - Fix Regex bug