+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unrelease]
-- Fix S3 Bucket forbidden
+## [3.6.0] - 2021-06-08
+### Added
 - Added #18389 - APM3.0 + Datahub3.0
 - Added #18718 - DataHub Helper
 - Added #19203 - Upgrade pandas version
+
+### Fixed
+- Fix S3 Bucket forbidden
+
 
 ## [3.4.2] - 2020-12-02
 - Fix Regex bug